Double类型改BigDecimal     private BigDecimal amount;     private BigDecimal tax;     private BigDecimal productPrice;

+++ cb-core/src/main/java/com/yunxin/cb/rb/entity/ReimbursementOrder.java
@@ -10,6 +10,7 @@ import org.hibernate.annotations.DynamicUpdate;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 
 import static javax.persistence.GenerationType.IDENTITY;
@@ -24,9 +25,9 @@ public class ReimbursementOrder implements java.io.Serializable {
     private int reimbursementOrderId;
     private OrderItem orderItem;
     private Reimbursement reimbursement;
-    private Double amount;
-    private Double tax;
-    private Double productPrice;
+    private BigDecimal amount;
+    private BigDecimal tax;
+    private BigDecimal productPrice;
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm")
     private Date createTime;
     @Id
@@ -78,27 +79,27 @@ public class ReimbursementOrder implements java.io.Serializable {
     }
 
     @Column(unique = true, nullable = false, insertable = true, updatable = true, length = 11)
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
     @Column(unique = true, nullable = false, insertable = true, updatable = true, length = 11)
-    public Double getTax() {
+    public BigDecimal getTax() {
         return tax;
     }
 
-    public void setTax(Double tax) {
+    public void setTax(BigDecimal tax) {
         this.tax = tax;
     }
     @Column(unique = true, nullable = false, insertable = true, updatable = true, length = 11)
-    public Double getProductPrice() {
+    public BigDecimal getProductPrice() {
         return productPrice;
     }
 
-    public void setProductPrice(Double productPrice) {
+    public void setProductPrice(BigDecimal productPrice) {
         this.productPrice = productPrice;
     }
 }
